# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiKeyCreate": {
        "description": "Request to create a new API key.",
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The key stops working after this time. Must be in the future.",
            "title": "Expires At"
          },
          "name": {
            "description": "Human-readable key name",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "organisation_id": {
            "description": "Organisation the key belongs to. The caller must be its organisation_manager.",
            "format": "uuid",
            "title": "Organisation Id",
            "type": "string"
          },
          "scopes": {
            "description": "Permission scopes (e.g. 'deployments:read')",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Scopes",
            "type": "array"
          }
        },
        "required": [
          "organisation_id",
          "name",
          "scopes"
        ],
        "title": "ApiKeyCreate",
        "type": "object"
      },
...
      "CreateDeploymentRequest": {
        "properties": {
          "deployment_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deployment End"
          },
          "deployment_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deployment Start"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "location_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Name"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "name"
        ],
        "title": "CreateDeploymentRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "An organisation API key, ww_live_...",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "V2 backend \u2014 async job system, model conversion",
    "title": "Wildlife Watcher API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/deployments": {
      "post": {
        "description": "Create a deployment in a project the caller can access (manual upload assignment).\n\nGuarded by project access \u2014 a caller with no role on the project gets 404 (so other tenants'\nproject IDs can't be probed). Auto-creates a placeholder device because deployments require a\nnon-null ``device_id`` (mirrors the CamtrapDP import); a fresh device id also satisfies the\n``(deployment_start, device_id)`` unique index.",
        "operationId": "create_deployment_api_deployments_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeploymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Deployment Api Deployments Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Deployment",
        "tags": [
          "deployments"
        ]
      }
    },
...
    "/api/jobs/{job_id}": {
      "get": {
        "description": "Poll the current status of an async job (scoped to the caller).",
        "operationId": "get_job_status_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Status",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/v1/api-keys": {
      "get": {
        "description": "List an organisation's active (unrevoked) API keys. Managers only.",
        "operationId": "list_keys_api_v1_api_keys_get",
        "parameters": [
          {
            "description": "Organisation whose keys to list",
            "in": "query",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "description": "Organisation whose keys to list",
              "format": "uuid",
              "title": "Organisation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Keys",
        "tags": [
          "public-api"
        ]
      },
      "post": {
        "description": "Create an API key for an organisation the caller manages.\n\nThe raw key is returned once. It cannot be retrieved again.",
        "operationId": "create_key_api_v1_api_keys_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Key",
        "tags": [
          "public-api"
        ]
      }
    },
    "/api/v1/api-keys/{key_id}": {
      "delete": {
        "description": "Revoke one of an organisation's API keys. Managers only.",
        "operationId": "revoke_key_api_v1_api_keys__key_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Key Id",
              "type": "string"
            }
          },
          {
            "description": "Organisation the key belongs to",
            "in": "query",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "description": "Organisation the key belongs to",
              "format": "uuid",
              "title": "Organisation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Key",
        "tags": [
          "public-api"
        ]
      }
    },
    "/api/v1/deployments": {
      "get": {
        "description": "List the deployments in the organisation's projects, newest first.",
        "operationId": "api_list_deployments_api_v1_deployments_get",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "description": "A deployment status value: planned, started or ended",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A deployment status value: planned, started or ended",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Api List Deployments",
        "tags": [
          "public-api"
        ]
      }
    },
    "/api/v1/deployments/{deployment_id}": {
      "get": {
        "description": "Get one deployment in the organisation's projects.",
        "operationId": "api_get_deployment_api_v1_deployments__deployment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Deployment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Api Get Deployment",
        "tags": [
          "public-api"
        ]
      }
    },
    "/api/v1/devices": {
      "get": {
        "description": "List the organisation's own cameras.",
        "operationId": "api_list_devices_api_v1_devices_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Api List Devices",
        "tags": [
          "public-api"
        ]
      }
    },
    "/api/v1/devices/{device_eui}/telemetry": {
      "get": {
        "description": "A camera's LoRaWAN messages from its deployments in the organisation's projects, newest first.",
        "operationId": "api_device_telemetry_api_v1_devices__device_eui__telemetry_get",
        "parameters": [
          {
            "in": "path",
            "name": "device_eui",
            "required": true,
            "schema": {
              "title": "Device Eui",
              "type": "string"
            }
          },
          {
            "description": "ISO 8601; messages received at or after",
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO 8601; messages received at or after",
              "title": "Date From"
            }
          },
          {
            "description": "ISO 8601; messages received at or before",
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO 8601; messages received at or before",
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Api Device Telemetry",
        "tags": [
          "public-api"
        ]
      }
    },
    "/api/v1/export/camtrapdp": {
      "post": {
        "description": "Start a CamtrapDP export of one project with its original photos, as the Download button does.\n\nReturns ``{job_id}``; poll ``GET /api/v1/jobs/{job_id}`` for the signed link. Behind\n``FF_CAMTRAPDP_EXPORT_ENABLED``, like the Download button's export.",
        "operationId": "api_export_camtrapdp_api_v1_export_camtrapdp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CamtrapExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Api Export Camtrapdp",
        "tags": [
          "public-api"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "description": "A job one of the organisation's keys started: status, progress, and the signed link once done.\n\nAnother organisation's job, or a signed-in user's, answers 404 like a missing one.",
        "operationId": "api_get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Api Get Job",
        "tags": [
          "public-api"
        ]
      }
    },
...
  }
}